Use the actual pool size instead of max pool size
The current pool size is probably more useful and interesting than a
(most likely) theoretical max pool size, should the two settings ever
diverge. If they were to change, it would probably make more sense to
include both the pool size and the max pool size again.

(Max pool size will only be used if the queue is full, according
to the ThreadPoolExecutor docs, and the queue size has to be set
explicitly.)